+++ libs/rxjs/utility/src/index.ts
@@ -8,10 +8,11 @@
  */
 /* istanbul ignore file */
 export { debounceWithQuery } from './lib/debounce-with-query';
+export { fromReadableStream } from './lib/from-readable-stream';
+export { fromFetchWithProgress } from './lib/from-fetch-with-progress';
 export { mapIfSource } from './lib/map-if-source';
+export { switchMapIfSource } from './lib/switch-map-if-source';
 export { tapIf } from './lib/tap-if';
 export { tapOnFirstEmit } from './lib/tap-on-first-emit';
 export { tapOnSubscribe } from './lib/tap-on-subscribe';
 export { toWritableStream } from './lib/to-writable-stream';
-export { fromReadableStream } from './lib/from-readable-stream';
-export { fromFetchWithProgress } from './lib/from-fetch-with-progress';